requirements.txt
Django==2.2.16
django-ckeditor
requests==2.26.0
